test: moved over all carousel test to MovieCarousel.test
